Assignment5/requirements.txt
pydoc 
string
tkinter.font 
pandas
urllib.request
ssl
os
re
requests
nltk
bs4
nltk.corpus
sklearn.feature_extraction.text
numpy
flask 
render_template
request
redirect
pymssql